Refactor and add language parameter handling in move_dots and related tests
