+++ src/index.php
@@ -17,8 +17,8 @@ use function WpRefs\ES\fix_es;
 use function WpRefs\ES\es_section;
 // use function WpRefs\DelDuplicateRefs\fix_refs_names;
 use function WpRefs\DelDuplicateRefs\remove_Duplicate_refs_With_attrs;
-use function WpRefs\MoveDots\move_dots_text;
-use function WpRefs\MoveDots\add_lang_en;
+use function WpRefs\MovesDots\move_dots_text;
+use function WpRefs\EnLangParam\add_lang_en;
 use function WpRefs\MdCat\add_Translated_from_MDWiki;
 use function WpRefs\Bots\Mini\mini_fixes;
 use function WpRefs\Bots\Mini\mini_fixes_after_fixing;
